fix(wextra): fix asan-build -Wsign-compare in multihost fabric tests
Fix sign-compare warnings in multihost fabric and ttswitch tests:
- multihost_fabric_fixtures.hpp: cast size() to int in system_supported()
- intermesh_routing.cpp: uint32_t loop var, cast rank() comparisons
- intermesh_routing_test_utils.cpp: cast rank() comparison
- socket_send_recv.cpp: size_t config_idx, cast rank() comparisons
- socket_send_recv_utils.cpp: uint32_t/size_t loop vars (13 loops)
- mesh_socket_test_context.cpp: uint32_t world_size, cast size() in loop
- test_mesh_device_ttswitch.cpp: uint32_t mesh_id_val (2 occurrences)

+++ tests/tt_metal/multihost/fabric_tests/multihost_fabric_fixtures.hpp
@@ -122,7 +122,7 @@ public:
         const auto& cluster = tt::tt_metal::MetalContext::instance().get_cluster();
         const auto& eth_coord_mapping = this->get_eth_coord_mapping();
         return *(tt::tt_metal::MetalContext::instance().global_distributed_context().size()) ==
-                   eth_coord_mapping.size() &&
+                   static_cast<int>(eth_coord_mapping.size()) &&
                cluster.user_exposed_chip_ids().size() == eth_coord_mapping[0].size();
     }
 };
@@ -153,7 +153,7 @@ public:
         const auto& cluster = tt::tt_metal::MetalContext::instance().get_cluster();
         const auto& eth_coord_mapping = this->get_eth_coord_mapping();
         return *(tt::tt_metal::MetalContext::instance().global_distributed_context().size()) ==
-                   eth_coord_mapping.size() &&
+                   static_cast<int>(eth_coord_mapping.size()) &&
                cluster.user_exposed_chip_ids().size() == eth_coord_mapping[0].size();
     }
 };
@@ -310,7 +310,7 @@ public:
         const auto& cluster = tt::tt_metal::MetalContext::instance().get_cluster();
         const auto& mesh_graph = tt::tt_metal::MetalContext::instance().get_control_plane().get_mesh_graph();
         return *(tt::tt_metal::MetalContext::instance().global_distributed_context().size()) ==
-                   mesh_graph.get_mesh_ids().size() &&
+                   static_cast<int>(mesh_graph.get_mesh_ids().size()) &&
                cluster.get_board_type(0) == BoardType::UBB;
     }
 };
@@ -335,7 +335,7 @@ public:
         const auto& cluster = tt::tt_metal::MetalContext::instance().get_cluster();
         const auto& mesh_graph = tt::tt_metal::MetalContext::instance().get_control_plane().get_mesh_graph();
         return *(tt::tt_metal::MetalContext::instance().global_distributed_context().size()) ==
-                   mesh_graph.get_mesh_ids().size() &&
+                   static_cast<int>(mesh_graph.get_mesh_ids().size()) &&
                cluster.is_ubb_galaxy();
     }
 };
